+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-### Release date ??? - Release ?.?.?
+### ??? - ??? Release date 
+
+### 25.10.2024 - Release 1.0.5
 
 -   [[#70]](https://github.com/Marcin-Migdal/m-component-library/issues/70) **[PATCH]** Export hslToHex helper
 -   [[#72]](https://github.com/Marcin-Migdal/m-component-library/issues/72) **[PATCH]** Dropdown improvements